Major refactor to allow testing of all 3 different types of input.
There is also a new Class that will eventually replace the
PressionEnthalpy class, which is Fonction. This new class is much more
clean and concise that is why we are still doint individual testing
from real world samples before we apply it to the original input file
that was provided.